pyproject.toml
[project]
name = "pagodo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4==4.13.5",
    "pyyaml>=6.0.3",
    "requests==2.32.5",
    "typer>=0.21.1",
]
[project.scripts]
pagodo = "pagodo.cli:app"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

